backend/requirements.txt
# GradeOps — Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9        # Form(...) / File(...) uploads

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9         # PostgreSQL driver
python-dotenv>=1.0.0           # .env file support

# ── PDF Processing ────────────────────────────────────────────────────────────
Pillow>=10.0.0
PyMuPDF>=1.24.0                # PDF text extraction (fitz)

# ── Remote LLM ────────────────────────────────────────────────────────────────
httpx>=0.27.0                  # Async HTTP client for HF Inference API
pydantic>=2.0.0